Protect against sign out or deletion of bootstrap admin.
