+++ src/main/java/org/sagebionetworks/bridge/user/TestUser.java
@@ -1,0 +1,53 @@
+package org.sagebionetworks.bridge.user;
+
+import java.io.IOException;
+import java.util.List;
+
+import org.sagebionetworks.bridge.rest.ClientManager;
+import org.sagebionetworks.bridge.rest.Config;
+import org.sagebionetworks.bridge.rest.model.Phone;
+import org.sagebionetworks.bridge.rest.model.Role;
+import org.sagebionetworks.bridge.rest.model.SignIn;
+import org.sagebionetworks.bridge.rest.model.UserSessionInfo;
+
+/**
+ * A utility class that holds a test account on Bridge. It provides an easy means to delete 
+ * the account at the end of the test. Generated by the helper methods of the TestUserHelper 
+ * class.
+ */
+public interface TestUser {
+    UserSessionInfo getSession();
+    String getEmail();
+    Phone getPhone();
+    String getPassword();
+    List<Role> getRoles();
+    String getDefaultSubpopulation();
+    String getAppId();
+    String getUserId();
+    
+    SignIn getSignIn();
+    Config getConfig();
+    ClientManager getClientManager();
+    
+    /** 
+     * Sign the account in again (whether currently signed in or not). 
+     *  
+     * @throws UnsupportedOperationException if the account is the bootstrap admin
+     */
+    UserSessionInfo signInAgain();
+    /**
+     * Sign the account out (whether currently signed in or not).
+     * 
+     * @throws UnsupportedOperationException if the account is the bootstrap admin
+     */
+    void signOut() throws IOException;
+    /**
+     * Sign the account out (whether currently signed in or not) and then delete it using
+     * the bootstrap admin account.
+     * 
+     * @throws UnsupportedOperationException if the account is the bootstrap admin itself
+     */
+    void signOutAndDeleteUser() throws IOException;
+
+    <T> T getClient(Class<T> service);
+}
